+++ Models/repositories/SemestersRepository.cs
@@ -6,11 +6,11 @@ namespace Web_API_for_scheduling.Models.repositories;
 public class SemestersRepository(TimetableDbContext context) : IRepository<Semester>
 {
     private readonly TimetableDbContext _context = context;
-    public async Task<bool> DeleteAsync(Guid id)
+    public async Task<bool?> DeleteAsync(Guid id)
     {
         var semester = await _context.Semester.FindAsync(id);
         if (semester == null) return false;
-
+        if (await _context.Week.FirstAsync(i => i.SemesterID == id) != null) return null;
         _context.Semester.Remove(semester);
         await _context.SaveChangesAsync();
         return true;
