+++ Models/repositories/WeeksRepository.cs
@@ -7,11 +7,11 @@ namespace Web_API_for_scheduling.Models.repositories
     public class WeeksRepository(TimetableDbContext context) : IRepository<Week>
     {
         private readonly TimetableDbContext _context = context;
-        public async Task<bool> DeleteAsync(Guid id)
+        public async Task<bool?> DeleteAsync(Guid id)
         {
             var week = await _context.Week.FindAsync(id);
             if (week == null) return false;
-
+            if (await _context.Day.FirstAsync(i => i.WeekID == id) != null) return null;
             _context.Week.Remove(week);
             await _context.SaveChangesAsync();
             return true;
